✨Feat: bookmark  에러 처리 완료

+++ src/bookmarks/bookmarks.controller.ts
@@ -12,6 +12,8 @@ import {
   Get,
   BadRequestException,
   Req,
+  InternalServerErrorException,
+  UnauthorizedException,
 } from "@nestjs/common";
 import { BookmarksService } from "./bookmarks.service";
 
@@ -27,63 +29,120 @@ import { OptionalAuthGuard } from "src/auth/authentication/guards/optionAuth.gua
 export class BookmarksController {
   constructor(private readonly bookmarksService: BookmarksService) {}
 
+  //북마크 토글
   @UseGuards(JwtAuthGuard)
   @Post("toggle")
   toggleBookmark(
     @Request() req: any,
     @Body(ValidationPipe) toggleBookmarkDto: ToggleBookmarkDto,
   ): Promise<ResponseBookmarkDto | MessageResponseDto> {
-    const userId = req.user.id;
-    const landmarkId = toggleBookmarkDto.landmarkId;
-    return this.bookmarksService.toggleBookmark(userId, landmarkId);
+    try {
+      const userId = req.user.id;
+      if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+      const landmarkId = toggleBookmarkDto.landmarkId;
+      if(!landmarkId) throw new BadRequestException("북마크에 추가되지 않는 랜드마크입니다.")
+
+      //토글
+      return this.bookmarksService.toggleBookmark(userId, landmarkId);
+    } 
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
   
+  //북마크 페이지
   @Get("/bookmarks")
   @UseGuards(OptionalAuthGuard)
   async getUserBookmarks(@Request() req: any): Promise<BookmarklistDto[]> {
-    console.log("req.user: ", req.user);
-    const userId = req.user?.id;
-    return this.bookmarksService.getBookmarksByUserId(userId);
+    try {
+      const userId = req.user.id;
+      if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+      return this.bookmarksService.getBookmarksByUserId(userId);
+      }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 
   //지역구별 리스트
   @Get("/user")
   @UseGuards(JwtAuthGuard)
   async findBookmarksByUser(@Request() req: any): Promise<SiDoBookmarkListDto[]> {
+    try {
     const userId = req.user.id;
+    if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
     return this.bookmarksService.findBookmarksByUser(userId);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 
+  //실제 사용 x api
   @Get(":landmarkId")
   @UseGuards(JwtAuthGuard)
   async findOneByUserAndLandmark(
     @Request() req: any,
     @Param("landmarkId", ParseIntPipe) landmarkId: number,
   ): Promise<ResponseBookmarkDto> {
-    const userId = req.user.id;
-    return this.bookmarksService.findOneByUserAndLandmark(userId, landmarkId);
-  }
+    try {
+      const userId = req.user.id;
+      if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
 
-  //유저의 랜드마크 아이디로 조회
+      if(!landmarkId) throw new BadRequestException("북마크에 추가되지 않는 랜드마크입니다.")
+
+      return this.bookmarksService.findOneByUserAndLandmark(userId, landmarkId);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
+  }
+  
+  //실제 사용 x api
   @Post()
   @UseGuards(JwtAuthGuard)
   async createBookmark(
     @Request() req: any,
     @Body(ValidationPipe) createBookmarkDto: CreateBookmarkDto,
   ): Promise<ResponseBookmarkDto> {
-    const userId = req.user.id; // 로그인된 사용자의 ID
-    const landmarkId = createBookmarkDto.landmarkId;
-    return this.bookmarksService.create(userId, landmarkId);
+    try {
+      const userId = req.user.id; // 로그인된 사용자의 ID
+      if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+      const landmarkId = createBookmarkDto.landmarkId;
+      if(!landmarkId) throw new BadRequestException("북마크에 추가되지 않는 랜드마크입니다.")
+
+      //db에 추가
+      return this.bookmarksService.create(userId, landmarkId);
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 
+  //북마크 삭제 (실제 사용 x)
   @Delete(":landmarkId")
   @UseGuards(JwtAuthGuard)
   async deleteBookmark(
     @Request() req: any,
     @Param("landmarkId", ParseIntPipe) landmarkId: number,
   ): Promise<void | MessageResponseDto> {
-    const userId = req.user.id; // 로그인된 사용자의 ID
-    const message = this.bookmarksService.delete(userId, landmarkId);
-    return message;
+    try {
+      if(!landmarkId) throw new BadRequestException("북마크에 추가되지 않는 랜드마크입니다.")
+
+      const userId = req.user.id; // 로그인된 사용자의 ID
+      if(!userId) throw new UnauthorizedException("로그인 되지 않은 사용자입니다.");
+
+      //db에서 삭제
+      const message = this.bookmarksService.delete(userId, landmarkId);
+      
+      return message;
+    }
+    catch (error) {
+      throw new InternalServerErrorException(`서버 오류 발생 : ${error.message}`)
+    }
   }
 }
